requirements.txt
django==2.2.5
# TODO: Change this to psycopg2-binary
psycopg2==2.7.4 
djangorestframework==3.12.1
Markdown==3.3.1
django-filter==2.4.0
